Fixed all warnings except a few in Windows that are worth investigating further.

+++ antlrcpp/CPPUtils.cpp
@@ -51,7 +51,7 @@ namespace antlrcpp {
     
     std::map<std::wstring, int>* toMap(const std::vector<std::wstring> &keys) {
         std::map<std::wstring, int>* m = new std::map<std::wstring, int>();
-        for (int i = 0; i < keys.size(); ++i) {
+        for (int i = 0; i < (int)keys.size(); ++i) {
             m->insert(std::pair<std::wstring, int>(keys[i], i));
         }
         return m;
@@ -76,7 +76,7 @@ namespace antlrcpp {
 		if (data == nullptr) return nullptr;
 		wchar_t* cdata = new wchar_t[data->size()];
 
-		for (int i = 0; i < data->size(); i++){
+		for (int i = 0; i < (int)data->size(); i++){
 			cdata[i] = (char)data->at(i);
 		}
